requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
pandas==1.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
wget==3.2
